requirements.txt
transformers
torch
datasets
sentence-transformers
faiss-cpu
pypdf
gradio